requirements.txt
gradio>=4.0.0
openai-whisper
yt-dlp
ffmpeg-python
